\name{corrfuns-package}
\alias{corrfuns-package}
\docType{package}
\title{
Correlation Coefficient Related Functions
}

\description{
Description: Many correlation coefficient related functions,
such as partial correlations and hypothesis testing.
}

\details{
\tabular{ll}{
Package: \tab corrfuns\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2023-10-26\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

%\references{
%}





